Set up directory structure. Created Signup component. Installed react-router-dom. Set up root route and signup route.
