+++ frontend/src/components/Signup/Signup.jsx
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import './Signup.scss';
+
+class Signup extends Component {
+
+  render() {
+    return (
+      <div className="signup">
+        This is a singup form
+      </div>
+    );
+  }
+}
+
+export default Signup;
