+++ frontend/src/components/Signup/index.js
@@ -1,0 +1,2 @@
+import Signup from './Signup';
+export default Signup;
